Move pkg/shipper code into ingester/shipper.go.
