+++ pkg/ingester/shipper_test.go
@@ -11,6 +11,8 @@ import (
 	"io"
 	"os"
 	"path"
+	"path/filepath"
+	"sort"
 	"testing"
 
 	"github.com/go-kit/log"
@@ -206,4 +208,116 @@ func TestShipper_DeceivingUploadErrors(t *testing.T) {
 	uploaded, err = s.Sync(context.Background())
 	require.NoError(t, err)
 	require.Equal(t, 1, uploaded)
+}
+
+func TestIterBlockMetas(t *testing.T) {
+	dir := t.TempDir()
+
+	id1 := ulid.MustNew(1, nil)
+	require.NoError(t, os.Mkdir(path.Join(dir, id1.String()), os.ModePerm))
+	require.NoError(t, metadata.Meta{
+		BlockMeta: tsdb.BlockMeta{
+			ULID:    id1,
+			MaxTime: 2000,
+			MinTime: 1000,
+			Version: 1,
+		},
+	}.WriteToDir(log.NewNopLogger(), path.Join(dir, id1.String())))
+
+	id2 := ulid.MustNew(2, nil)
+	require.NoError(t, os.Mkdir(path.Join(dir, id2.String()), os.ModePerm))
+	require.NoError(t, metadata.Meta{
+		BlockMeta: tsdb.BlockMeta{
+			ULID:    id2,
+			MaxTime: 5000,
+			MinTime: 4000,
+			Version: 1,
+		},
+	}.WriteToDir(log.NewNopLogger(), path.Join(dir, id2.String())))
+
+	id3 := ulid.MustNew(3, nil)
+	require.NoError(t, os.Mkdir(path.Join(dir, id3.String()), os.ModePerm))
+	require.NoError(t, metadata.Meta{
+		BlockMeta: tsdb.BlockMeta{
+			ULID:    id3,
+			MaxTime: 3000,
+			MinTime: 2000,
+			Version: 1,
+		},
+	}.WriteToDir(log.NewNopLogger(), path.Join(dir, id3.String())))
+
+	shipper := NewShipper(nil, nil, dir, nil, metadata.TestSource, metadata.NoneFunc)
+	metas, err := shipper.blockMetasFromOldest()
+	require.NoError(t, err)
+	require.Equal(t, sort.SliceIsSorted(metas, func(i, j int) bool {
+		return metas[i].BlockMeta.MinTime < metas[j].BlockMeta.MinTime
+	}), true)
+}
+
+func TestShipperAddsSegmentFiles(t *testing.T) {
+	dir := t.TempDir()
+
+	inmemory := objstore.NewInMemBucket()
+
+	s := NewShipper(nil, nil, dir, inmemory, metadata.TestSource, metadata.NoneFunc)
+
+	id := ulid.MustNew(1, nil)
+	blockDir := path.Join(dir, id.String())
+	chunksDir := path.Join(blockDir, block.ChunksDirname)
+	require.NoError(t, os.MkdirAll(chunksDir, os.ModePerm))
+
+	// Prepare minimal "block" for shipper (meta.json, index, one segment file).
+	require.NoError(t, metadata.Meta{
+		BlockMeta: tsdb.BlockMeta{
+			ULID:    id,
+			MaxTime: 2000,
+			MinTime: 1000,
+			Version: 1,
+			Stats: tsdb.BlockStats{
+				NumSamples: 1000, // Not really, but shipper needs nonzero value.
+			},
+		},
+	}.WriteToDir(log.NewNopLogger(), path.Join(dir, id.String())))
+	require.NoError(t, os.WriteFile(filepath.Join(blockDir, "index"), []byte("index file"), 0666))
+	segmentFile := "00001"
+	require.NoError(t, os.WriteFile(filepath.Join(chunksDir, segmentFile), []byte("hello world"), 0666))
+
+	uploaded, err := s.Sync(context.Background())
+	require.NoError(t, err)
+	require.Equal(t, 1, uploaded)
+
+	meta, err := block.DownloadMeta(context.Background(), log.NewNopLogger(), inmemory, id)
+	require.NoError(t, err)
+
+	require.Equal(t, []string{segmentFile}, meta.Thanos.SegmentFiles)
+}
+
+func TestReadMetaFile(t *testing.T) {
+	t.Run("Missing meta file", func(t *testing.T) {
+		// Create TSDB directory without meta file
+		dpath := t.TempDir()
+
+		_, err := readShipperMetaFile(dpath)
+		fpath := filepath.Join(dpath, shipperMetaFilename)
+		require.Equal(t, fmt.Sprintf(`failed to read %s: open %s: no such file or directory`, fpath, fpath), err.Error())
+	})
+
+	t.Run("Non-JSON meta file", func(t *testing.T) {
+		dpath := t.TempDir()
+		fpath := filepath.Join(dpath, shipperMetaFilename)
+		// Make an invalid JSON file
+		require.NoError(t, os.WriteFile(fpath, []byte("{"), 0600))
+
+		_, err := readShipperMetaFile(dpath)
+		require.Equal(t, fmt.Sprintf(`failed to parse %s as JSON: "{": unexpected end of JSON input`, fpath), err.Error())
+	})
+
+	t.Run("Wrongly versioned meta file", func(t *testing.T) {
+		dpath := t.TempDir()
+		fpath := filepath.Join(dpath, shipperMetaFilename)
+		require.NoError(t, os.WriteFile(fpath, []byte(`{"version": 2}`), 0600))
+
+		_, err := readShipperMetaFile(dpath)
+		require.Equal(t, "unexpected meta file version 2", err.Error())
+	})
 }
